Move IsAllFinite from training to contrib op
