+++ orttraining/orttraining/training_ops/cuda/math/isfinite.h
@@ -21,32 +21,5 @@ class IsFiniteOp final : public CudaKernel {
 template <typename TSrc>
 void IsFinite(cudaStream_t stream, const TSrc* input, bool* output, size_t N);
 
-template <typename TSrc>
-class IsAllFiniteOp final : public CudaKernel {
- public:
-  IsAllFiniteOp(const OpKernelInfo& info) : CudaKernel(info) {
-    int64_t isinf_only;
-    info.GetAttrOrDefault("isinf_only", &isinf_only, static_cast<int64_t>(0));
-    isinf_only_ = (isinf_only != 0);
-
-    int64_t isnan_only;
-    info.GetAttrOrDefault("isnan_only", &isnan_only, static_cast<int64_t>(0));
-    isnan_only_ = (isnan_only != 0);
-
-    ORT_ENFORCE(!(isinf_only_ && isnan_only_),
-                "Both attributes isinf_only and isnan_only cannot be set. Unset both to check for both conditions.");
-  }
-
-  Status ComputeInternal(OpKernelContext* context) const override;
-
-  private:
-  bool isinf_only_, isnan_only_;
-};
-
-template <typename T>
-struct IsAllFiniteFunctor {
-  void operator()(cudaStream_t stream, ChunkGroup<1> chunks, bool* output, const bool isinf_only, const bool isnan_only);
-};
-
 }  // namespace cuda
 }  // namespace onnxruntime
